@@ -4,58 +4,31 @@ declare(strict_types=1);
 
 namespace JWage\PHPUnitTestGenerator\Configuration;
 
-class Configuration
+final readonly class Configuration
 {
-    /** @var string */
-    private $autoloadingStrategy;
+    public function __construct(private string $autoloadingStrategy, private string $sourceNamespace, private string $sourceDir, private string $testsNamespace, private string $testsDir) {}
 
-    /** @var string */
-    private $sourceNamespace;
-
-    /** @var string */
-    private $sourceDir;
-
-    /** @var string */
-    private $testsNamespace;
-
-    /** @var string */
-    private $testsDir;
-
-    public function __construct(
-        string $autoloadingStrategy,
-        string $sourceNamespace,
-        string $sourceDir,
-        string $testsNamespace,
-        string $testsDir
-    ) {
-        $this->autoloadingStrategy = $autoloadingStrategy;
-        $this->sourceNamespace     = $sourceNamespace;
-        $this->sourceDir           = $sourceDir;
-        $this->testsNamespace      = $testsNamespace;
-        $this->testsDir            = $testsDir;
-    }
-
-    public function getAutoloadingStrategy() : string
+    public function getAutoloadingStrategy(): string
     {
         return $this->autoloadingStrategy;
     }
 
-    public function getSourceNamespace() : string
+    public function getSourceNamespace(): string
     {
         return $this->sourceNamespace;
     }
 
-    public function getSourceDir() : string
+    public function getSourceDir(): string
     {
         return $this->sourceDir;
     }
 
-    public function getTestsNamespace() : string
+    public function getTestsNamespace(): string
     {
         return $this->testsNamespace;
     }
 
-    public function getTestsDir() : string
+    public function getTestsDir(): string
     {
         return $this->testsDir;
     }
