@@ -11,22 +11,16 @@ use JWage\PHPUnitTestGenerator\Configuration\Configuration;
 use JWage\PHPUnitTestGenerator\TestClassGenerator;
 use JWage\PHPUnitTestGenerator\Writer\Psr4TestClassWriter;
 use JWage\PHPUnitTestGenerator\Writer\TestClassWriter;
+use Override;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use function array_diff;
-use function array_values;
-use function assert;
-use function file_exists;
-use function get_declared_classes;
-use function getcwd;
-use function is_string;
-use function sprintf;
 
-class GenerateTestClassCommand extends Command
+final class GenerateTestClassCommand extends Command
 {
-    protected function configure() : void
+    #[Override]
+    protected function configure(): void
     {
         $this
             ->setName('generate-test-class')
@@ -34,7 +28,8 @@ class GenerateTestClassCommand extends Command
             ->addArgument('class', InputArgument::OPTIONAL, 'The class name to generate the test for.');
     }
 
-    protected function execute(InputInterface $input, OutputInterface $output) : void
+    #[Override]
+    protected function execute(InputInterface $input, OutputInterface $output): void
     {
         $className = $input->getArgument('class');
 
@@ -42,11 +37,11 @@ class GenerateTestClassCommand extends Command
             throw new InvalidArgumentException('Specify class name to generate a unit test for.');
         }
 
-        assert(is_string($className));
+        \assert(\is_string($className));
 
         $className = $this->getClassName($className);
 
-        $output->writeln(sprintf('Generating test class for <info>%s</info>', $className));
+        $output->writeln(\sprintf('Generating test class for <info>%s</info>', $className));
         $output->writeln('');
 
         $configuration = $this->createConfiguration();
@@ -60,25 +55,25 @@ class GenerateTestClassCommand extends Command
         $writePath = $this->createTestClassWriter($configuration)
             ->write($generatedTestClass);
 
-        $output->writeln(sprintf('Test class written to <info>%s</info>', $writePath));
+        $output->writeln(\sprintf('Test class written to <info>%s</info>', $writePath));
     }
 
-    private function getClassName(string $className) : string
+    private function getClassName(string $className): string
     {
         // path to class was given
-        $filePath = getcwd() . '/' . $className;
+        $filePath = \getcwd() . '/' . $className;
 
-        if (file_exists($filePath)) {
-            $beforeClasses = get_declared_classes();
+        if (\file_exists($filePath)) {
+            $beforeClasses = \get_declared_classes();
 
             require_once $filePath;
 
-            $afterClasses = get_declared_classes();
+            $afterClasses = \get_declared_classes();
 
-            $newClasses = array_values(array_diff($afterClasses, $beforeClasses));
+            $newClasses = \array_values(\array_diff($afterClasses, $beforeClasses));
 
-            if (! isset($newClasses[0])) {
-                throw new InvalidArgumentException(sprintf('Could not find class in file %s', $filePath));
+            if ( ! isset($newClasses[0])) {
+                throw new InvalidArgumentException(\sprintf('Could not find class in file %s', $filePath));
             }
 
             $className = $newClasses[0];
@@ -87,17 +82,17 @@ class GenerateTestClassCommand extends Command
         return $className;
     }
 
-    private function createConfiguration() : Configuration
+    private function createConfiguration(): Configuration
     {
         return (new ComposerConfigurationReader())->createConfiguration();
     }
 
-    private function createTestClassGenerator(Configuration $configuration) : TestClassGenerator
+    private function createTestClassGenerator(Configuration $configuration): TestClassGenerator
     {
         return new TestClassGenerator($configuration);
     }
 
-    private function createTestClassWriter(Configuration $configuration) : TestClassWriter
+    private function createTestClassWriter(Configuration $configuration): TestClassWriter
     {
         $autoloadingStrategy = $configuration->getAutoloadingStrategy();
 
@@ -106,7 +101,7 @@ class GenerateTestClassCommand extends Command
         }
 
         throw new InvalidArgumentException(
-            sprintf('Autoloading strategy not supported %s not supported', $autoloadingStrategy)
+            \sprintf('Autoloading strategy not supported %s not supported', $autoloadingStrategy),
         );
     }
 }
