@@ -6,59 +6,32 @@ namespace JWage\PHPUnitTestGenerator\Tests;
 
 class TestClass1
 {
-    /** @var TestDependency */
-    private $testDependency;
-
-    /** @var float */
-    private $testFloatArgument;
-
-    /** @var int */
-    private $testIntegerArgument;
-
-    /** @var string */
-    private $testStringArgument;
-
-    /** @var mixed[] */
-    private $testArrayArgument;
-
     /**
      * @param mixed[] $testArrayArgument
      */
-    public function __construct(
-        TestDependency $testDependency,
-        float $testFloatArgument,
-        int $testIntegerArgument,
-        string $testStringArgument,
-        array $testArrayArgument
-    ) {
-        $this->testDependency      = $testDependency;
-        $this->testFloatArgument   = $testFloatArgument;
-        $this->testIntegerArgument = $testIntegerArgument;
-        $this->testStringArgument  = $testStringArgument;
-        $this->testArrayArgument   = $testArrayArgument;
-    }
+    public function __construct(private TestDependency $testDependency, private readonly float $testFloatArgument, private readonly int $testIntegerArgument, private readonly string $testStringArgument, private readonly array $testArrayArgument) {}
 
-    public function getTestDependency() : TestDependency
+    public function getTestDependency(): TestDependency
     {
         return $this->testDependency;
     }
 
-    public function setTestDependency(TestDependency $testDependency) : void
+    public function setTestDependency(TestDependency $testDependency): void
     {
         $this->testDependency = $testDependency;
     }
 
-    public function getTestFloatArgument() : float
+    public function getTestFloatArgument(): float
     {
         return $this->testFloatArgument;
     }
 
-    public function getTestIntegerArgument() : int
+    public function getTestIntegerArgument(): int
     {
         return $this->testIntegerArgument;
     }
 
-    public function getTestStringArgument() : string
+    public function getTestStringArgument(): string
     {
         return $this->testStringArgument;
     }
@@ -66,21 +39,19 @@ class TestClass1
     /**
      * @return mixed[]
      */
-    public function getTestArrayArgument() : array
+    public function getTestArrayArgument(): array
     {
         return $this->testArrayArgument;
     }
 
-    public function getTestMethodWithArguments(string $a, float $b, int $c) : void
-    {
-    }
+    public function getTestMethodWithArguments(string $a, float $b, int $c): void {}
 
-    public function getSomething() : string
+    public function getSomething(): string
     {
         return 'something';
     }
 
-    public function getTestBoolean() : bool
+    public function getTestBoolean(): bool
     {
         return true;
     }
@@ -88,7 +59,7 @@ class TestClass1
     /**
      * @return mixed[]
      */
-    public function getTestArray() : array
+    public function getTestArray(): array
     {
         return [];
     }
