@@ -4,6 +4,4 @@ declare(strict_types=1);
 
 namespace JWage\PHPUnitTestGenerator\Tests;
 
-class TestDependency
-{
-}
+class TestDependency {}
